Last OOB effect for EQS
Also removed 00_countries because it was extraneous.

Added effects for solarwar.7, as well as renamed some of the advisor GFX files for easier sorting. Added the vanilla triggers file so that we can implement what nations do and don't have custom designers and advisors, and also added in a random names list for them, too.