Decisions and tree change for Aurora
No more mandatory temple construction. Will be a decision branch like STG's Supreme Soviet building, with the same effects as the focuses once completed. Added focus and events for Nebula Flare to be captured and either executed or turned back to the Solar Empire.

Decisions for Aurora

Now there will be a series of decisions for constructing the temple, and it is no longer required for Aurora to construct it.

Added loc and opinion modifiers to CER, NER, and CLD for their spirits. Opinion modifiers are for CER and CLD, as they do not see each other as successors to EQS.